requirements.txt
python-telegram-bot
groq
apscheduler